Use environment.d and .profile for setup
There is no clear way to set environment variables before user session starts.
`~/.pam_environment` used to be a way to do so, but it was deprecated.
It was not a big issue for Zsh because `~/.zshenv` could be used to bootstrap environment variables for XDG Base Directory.
However, Fish uses XDG Base Directory to find configuration, meaning `$XDG_CONFIG_DIR` must be set before Fish starts.

To break this chicken-and-egg problem, the default shell is set to POSIX-compliant shell, which can set XDG Base Directory environment variables in `~/.profile`.
For any interactive usage, Fish is `exec`ed from the default shell.

Additionally, the shell profile loads from `environment.d` to ensure the same environment variables are available for both user sessions and user Systemd services.