Update Neovim configs
The background option is set only if the background color can be detected.
It has default of "dark".
The background option is explicitly set to "light" because I only use light themes.

The default values for the fileformats option is usually correct.
The explicit preference for unix line ending causes problems when working in Windows.